Server/requirements.txt
grpcio
grpcio-tools
SQLAlchemy
psycopg2-binary
paho-mqtt